upstream update

+++ Dev/Dev2.Studio.Core/Workspaces/WorkspaceItemRepository.cs
@@ -1,7 +1,6 @@
-
 /*
-*  Warewolf - The Easy Service Bus
-*  Copyright 2015 by Warewolf Ltd <alpha@warewolf.io>
+*  Warewolf - Once bitten, there's no going back
+*  Copyright 2018 by Warewolf Ltd <alpha@warewolf.io>
 *  Licensed under GNU Affero General Public License 3.0 or later. 
 *  Some rights reserved.
 *  Visit our website for more information <http://warewolf.io/>
@@ -14,27 +13,18 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml.Linq;
-using Dev2.Communication;
-using Dev2.Controller;
-using Dev2.Studio.Core.AppResources.Enums;
-using Dev2.Studio.Core.Interfaces;
 using Dev2.Studio.Core.Workspaces;
+using Dev2.Studio.Core;
+using Dev2.Studio.Interfaces;
+using Dev2.Studio.Interfaces.Enums;
+
 
 namespace Dev2.Workspaces
 {
     public class WorkspaceItemRepository : IWorkspaceItemRepository
     {
         #region Singleton Instance
-
-        //
-        // Multi-threaded implementation - see http://msdn.microsoft.com/en-us/library/ff650316.aspx
-        //
-        // This approach ensures that only one instance is created and only when the instance is needed. 
-        // Also, the variable is declared to be volatile to ensure that assignment to the instance variable
-        // completes before the instance variable can be accessed. Lastly, this approach uses a syncRoot 
-        // instance to lock on, rather than locking on the type itself, to avoid deadlocks.
-        //
-
+        
         static volatile IWorkspaceItemRepository _instance;
         static readonly object SyncRoot = new Object();
 
@@ -59,12 +49,9 @@ namespace Dev2.Workspaces
 
         #endregion
 
-        private IList<IWorkspaceItem> _workspaceItems;
+        IList<IWorkspaceItem> _workspaceItems;
 
-        public IList<IWorkspaceItem> WorkspaceItems
-        {
-            get { return _workspaceItems ?? (_workspaceItems = Read()); }
-        }
+        public IList<IWorkspaceItem> WorkspaceItems => _workspaceItems ?? (_workspaceItems = Read());
 
         #region CTOR
 
@@ -73,7 +60,6 @@ namespace Dev2.Workspaces
         {
         }
 
-        // BUG 9492 - 2013.06.08 - TWR : added constructor - use for testing only!
         public WorkspaceItemRepository(string repositoryPath)
         {
             _repositoryPath = repositoryPath;
@@ -81,14 +67,15 @@ namespace Dev2.Workspaces
 
         public WorkspaceItemRepository(IWorkspaceItemRepository workspaceItemRepository)
         {
+#pragma warning disable S3010 // For testing
             _instance = workspaceItemRepository;
+#pragma warning restore S3010
         }
 
         #endregion
 
         #region RepositoryPath
 
-        // BUG 9492 - 2013.06.08 - TWR : made public and non-static
         string _repositoryPath;
         public string RepositoryPath
         {
@@ -114,19 +101,19 @@ namespace Dev2.Workspaces
 
         #region Read
 
-        private IList<IWorkspaceItem> Read()
+        IList<IWorkspaceItem> Read()
         {
             var result = new List<IWorkspaceItem>();
-            if(File.Exists(RepositoryPath))
+            if (File.Exists(RepositoryPath))
             {
                 try
                 {
                     var xml = XElement.Parse(File.ReadAllText(RepositoryPath));
                     result.AddRange(xml.Elements().Select(x => new WorkspaceItem(x)));
                 }
-                // ReSharper disable EmptyGeneralCatchClause
+
                 catch
-                // ReSharper restore EmptyGeneralCatchClause
+
                 {
                     // corrupt so ignore
                 }
@@ -149,12 +136,12 @@ namespace Dev2.Workspaces
 
             if(!File.Exists(RepositoryPath))
             {
-                FileInfo fileInfo = new FileInfo(RepositoryPath);
-                if(fileInfo.Directory != null)
+                var fileInfo = new FileInfo(RepositoryPath);
+                if (fileInfo.Directory != null)
                 {
-                    string finalDirectoryPath = fileInfo.Directory.FullName;
+                    var finalDirectoryPath = fileInfo.Directory.FullName;
 
-                    if(!Directory.Exists(finalDirectoryPath))
+                    if (!Directory.Exists(finalDirectoryPath))
                     {
                         Directory.CreateDirectory(finalDirectoryPath);
                     }
@@ -173,14 +160,14 @@ namespace Dev2.Workspaces
             {
                 throw new ArgumentNullException("model");
             }
-            var workspaceItem = WorkspaceItems.FirstOrDefault(wi => wi.ID == model.ID && wi.EnvironmentID == model.Environment.ID);
+            var workspaceItem = WorkspaceItems.FirstOrDefault(wi => wi.ID == model.ID && wi.EnvironmentID == model.Environment.EnvironmentID);
             if(workspaceItem != null)
             {
                 return;
             }
 
             var context = model.Environment.Connection;
-            WorkspaceItems.Add(new WorkspaceItem(context.WorkspaceID, context.ServerID, model.Environment.ID, model.ID)
+            WorkspaceItems.Add(new WorkspaceItem(context.WorkspaceID, context.ServerID, model.Environment.EnvironmentID, model.ID)
             {
                 ServiceName = model.ResourceName,
                 IsWorkflowSaved = model.IsWorkflowSaved,
@@ -203,7 +190,7 @@ namespace Dev2.Workspaces
             {
                 throw new ArgumentNullException("resourceModel");
             }
-            var workspaceItem = WorkspaceItems.FirstOrDefault(wi => wi.ID == resourceModel.ID && wi.EnvironmentID == resourceModel.Environment.ID);
+            var workspaceItem = WorkspaceItems.FirstOrDefault(wi => wi.ID == resourceModel.ID && wi.EnvironmentID == resourceModel.Environment.EnvironmentID);
 
             if(workspaceItem == null)
             {
@@ -212,46 +199,12 @@ namespace Dev2.Workspaces
             workspaceItem.IsWorkflowSaved = resourceModel.IsWorkflowSaved;
         }
 
-        public ExecuteMessage UpdateWorkspaceItem(IContextualResourceModel resource, bool isLocalSave)
-        {
-            // BUG 9492 - 2013.06.08 - TWR : added null check
-            if(resource == null)
-            {
-                throw new ArgumentNullException("resource");
-            }
-            var workspaceItem = WorkspaceItems.FirstOrDefault(wi => wi.ID == resource.ID && wi.EnvironmentID == resource.Environment.ID);
-
-            if(workspaceItem == null)
-            {
-                var msg = new ExecuteMessage { HasError = false };
-                msg.SetMessage(string.Empty);
-                return msg;
-            }
-
-
-            workspaceItem.Action = WorkspaceItemAction.Commit;
-
-            var comsController = new CommunicationController { ServiceName = "UpdateWorkspaceItemService" };
-            comsController.AddPayloadArgument("Roles", String.Join(",", "Test"));
-            var xml = workspaceItem.ToXml();
-
-            comsController.AddPayloadArgument("ItemXml", xml.ToString(SaveOptions.DisableFormatting));
-            comsController.AddPayloadArgument("IsLocalSave", isLocalSave.ToString());
-
-            var con = resource.Environment.Connection;
-
-            var result = comsController.ExecuteCommand<ExecuteMessage>(con, con.WorkspaceID);
-
-            return result;
-        }
-
         #endregion
 
         #region Remove
 
         public void Remove(IContextualResourceModel resourceModel)
         {
-            // BUG 9492 - 2013.06.08 - TWR : added null check
             if(resourceModel == null)
             {
                 return;
@@ -264,6 +217,19 @@ namespace Dev2.Workspaces
 
             WorkspaceItems.Remove(itemToRemove);
             Write();
+            if (!resourceModel.IsNotWarewolfPath)
+            {
+                resourceModel.Environment.ResourceRepository.DeleteResourceFromWorkspaceAsync(resourceModel);
+            }
+        }
+
+        public void ClearWorkspaceItems(IContextualResourceModel resourceModel)
+        {
+            if (resourceModel == null)
+            {
+                return;
+            }
+            WorkspaceItems.Clear();
             resourceModel.Environment.ResourceRepository.DeleteResourceFromWorkspace(resourceModel);
         }
